# tests/m3/openapi/openapi_snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PSIT 高价值客户异常售后决策支持 API 合同",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/batches": {
      "post": {
        "summary": "Create Batch",
        "description": "上传一个标准 ZIP 新建批次；重复包 200 幂等命中（M3-04）。",
        "operationId": "create_batch_api_v1_batches_post",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_batch_api_v1_batches_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchWorkspaceView"
                }
              }
            }
          },
          "200": {
            "description": "重复包幂等命中",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchWorkspaceView"
                }
              }
            }
          },
          "400": {
            "description": "ANSWER_LEAKAGE_DETECTED：运行包发现答案或答案性质字段",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          },
          "409": {
            "description": "ACTIVE_BATCH_EXISTS：已有首次批量分析正在运行",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          },
          "413": {
            "description": "UPLOAD_TOO_LARGE：ZIP 超过冻结的上传上限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          },
          "415": {
            "description": "UNSUPPORTED_MEDIA_TYPE：上传或证据媒体类型不受支持",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          },
          "422": {
            "description": "REQUEST_VALIDATION_FAILED：HTTP 字段、联合请求、类型或枚举不合格",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR：未预期内部异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Batches",
        "description": "按 imported_at 倒序、batch_id 稳定排序列出批次（M3-04）。",
        "operationId": "list_batches_api_v1_batches_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchListView"
                }
              }
            }
          },
          "422": {
            "description": "REQUEST_VALIDATION_FAILED：HTTP 字段、联合请求、类型或枚举不合格",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR：未预期内部异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/batches/{batch_id}": {
      "get": {
        "summary": "Get Batch",
        "description": "读取单个批次工作台视图（M3-04）。",
        "operationId": "get_batch_api_v1_batches__batch_id__get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchWorkspaceView"
                }
              }
            }
          },
          "404": {
            "description": "RESOURCE_NOT_FOUND：批次、案例或证据不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          },
          "422": {
            "description": "REQUEST_VALIDATION_FAILED：HTTP 字段、联合请求、类型或枚举不合格",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR：未预期内部异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/batches/{batch_id}/runs": {
      "post": {
        "summary": "Start Batch Run",
        "description": "开始首次批量分析；无请求体，缺密钥时 503（M3-05 实现）。",
        "operationId": "start_batch_run_api_v1_batches__batch_id__runs_post",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchWorkspaceView"
                }
              }
            }
          },
          "404": {
            "description": "RESOURCE_NOT_FOUND：批次、案例或证据不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          },
          "409": {
            "description": "ACTIVE_BATCH_EXISTS：已有首次批量分析正在运行",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          },
          "503": {
            "description": "ANALYSIS_UNAVAILABLE：缺密钥、模型健康检查失败或应用正在正常关闭",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          },
          "422": {
            "description": "REQUEST_VALIDATION_FAILED：HTTP 字段、联合请求、类型或枚举不合格",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR：未预期内部异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/batches/{batch_id}/cases": {
      "get": {
        "summary": "List Cases",
        "description": "案例队列；服务端固定排序，不提供 sort_by（M3-04）。",
        "operationId": "list_cases_api_v1_batches__batch_id__cases_get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "只允许一个案例状态筛选",
              "title": "Status"
            },
            "description": "只允许一个案例状态筛选"
          },
          {
            "name": "intervention_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "只允许一个介入等级筛选",
              "title": "Intervention Level"
            },
            "description": "只允许一个介入等级筛选"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseQueueView"
                }
              }
            }
          },
          "404": {
            "description": "RESOURCE_NOT_FOUND：批次、案例或证据不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          },
          "422": {
            "description": "REQUEST_VALIDATION_FAILED：HTTP 字段、联合请求、类型或枚举不合格",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR：未预期内部异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/batches/{batch_id}/cases/{case_id}": {
      "get": {
        "summary": "Get Case",
        "description": "案例详情（M3-04）。",
        "operationId": "get_case_api_v1_batches__batch_id__cases__case_id__get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          },
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetailView"
                }
              }
            }
          },
          "404": {
            "description": "RESOURCE_NOT_FOUND：批次、案例或证据不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          },
          "422": {
            "description": "REQUEST_VALIDATION_FAILED：HTTP 字段、联合请求、类型或枚举不合格",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR：未预期内部异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/batches/{batch_id}/cases/{case_id}/reruns": {
      "post": {
        "summary": "Rerun Case",
        "description": "人工从头重跑单个案例；无请求体（M3-06 实现）。",
        "operationId": "rerun_case_api_v1_batches__batch_id__cases__case_id__reruns_post",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          },
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetailView"
                }
              }
            }
          },
          "404": {
            "description": "RESOURCE_NOT_FOUND：批次、案例或证据不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          },
          "409": {
            "description": "CASE_ALREADY_COMPLETED：已完成案例收到新的审核或重跑请求",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          },
          "503": {
            "description": "ANALYSIS_UNAVAILABLE：缺密钥、模型健康检查失败或应用正在正常关闭",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          },
          "422": {
            "description": "REQUEST_VALIDATION_FAILED：HTTP 字段、联合请求、类型或枚举不合格",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR：未预期内部异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/batches/{batch_id}/cases/{case_id}/reviews": {
      "post": {
        "summary": "Submit Review",
        "description": "提交人工确认；相同 submission_id 返回第一次保存结果（M3-07 实现）。",
        "operationId": "submit_review_api_v1_batches__batch_id__cases__case_id__reviews_post",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          },
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/ApprovedReviewRequest"
                  },
                  {
                    "$ref": "#/components/schemas/ModifiedApprovedReviewRequest"
                  },
                  {
                    "$ref": "#/components/schemas/RejectedWithJudgmentReviewRequest"
                  },
                  {
                    "$ref": "#/components/schemas/InsufficientEvidenceReviewRequest"
                  }
                ],
                "discriminator": {
                  "propertyName": "outcome",
                  "mapping": {
                    "APPROVED": "#/components/schemas/ApprovedReviewRequest",
                    "MODIFIED_AND_APPROVED": "#/components/schemas/ModifiedApprovedReviewRequest",
                    "REJECTED_WITH_JUDGMENT": "#/components/schemas/RejectedWithJudgmentReviewRequest",
                    "INSUFFICIENT_EVIDENCE": "#/components/schemas/InsufficientEvidenceReviewRequest"
                  }
                },
                "description": "四种人工确认请求之一",
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResultView"
                }
              }
            }
          },
          "200": {
            "description": "相同 submission_id 幂等命中",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResultView"
                }
              }
            }
          },
          "404": {
            "description": "RESOURCE_NOT_FOUND：批次、案例或证据不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          },
          "409": {
            "description": "CASE_ALREADY_COMPLETED：已完成案例收到新的审核或重跑请求",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          },
          "422": {
            "description": "REQUEST_VALIDATION_FAILED：HTTP 字段、联合请求、类型或枚举不合格",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR：未预期内部异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchWorkspaceView": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "source_filename": {
            "type": "string",
            "title": "Source Filename"
          },
          "is_mock": {
            "type": "boolean",
            "title": "Is Mock"
          },
          "status": {
            "$ref": "#/components/schemas/BatchStatus"
          },
          "case_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Case Count"
          },
          "evidence_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Evidence Count"
          },
          "analysis_succeeded_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Analysis Succeeded Count"
          },
          "error_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Error Count"
          },
          "imported_at": {
            "type": "string",
            "format": "date-time",
            "title": "Imported At"
          },
          "can_start_analysis": {
            "type": "boolean",
            "title": "Can Start Analysis"
          },
          "analysis_unavailable_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Unavailable Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "batch_id",
          "source_filename",
          "is_mock",
          "status",
          "case_count",
          "evidence_count",
          "analysis_succeeded_count",
          "error_count",
          "imported_at",
          "can_start_analysis"
        ],
        "title": "BatchWorkspaceView",
        "description": "批次工作台视图（规格 12.2）。"
      },
      "Body_create_batch_api_v1_batches_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "标准 ZIP 运行包"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_create_batch_api_v1_batches_post"
      },
...
      "BusinessError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "object_type": {
            "type": "string",
            "title": "Object Type"
          },
          "object_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Object Id"
          },
          "stage": {
            "$ref": "#/components/schemas/ProcessingErrorStage"
          },
          "next_action": {
            "type": "string",
            "title": "Next Action"
          },
          "trace_id": {
            "type": "string",
            "minLength": 1,
            "title": "Trace Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "object_type",
          "stage",
          "next_action"
        ],
        "title": "BusinessError",
        "description": "统一 JSON 错误结构（规格 12.4）。"
      },
...
    }
  }
}